+++ src/main/java/com/github/gather/dto/response/UserInfoResponse.java
@@ -5,17 +5,22 @@ import lombok.*;
 
 @Getter
 @Setter
-@AllArgsConstructor
-@NoArgsConstructor
 public class UserInfoResponse {
 
     private Long userId;
     private String nickname;
     private String email;
-    private String password;
     private String phoneNumber;
     private String image;
     private Location locationId;
 
+    public UserInfoResponse(Long userId, String nickname, String email, String phoneNumber, String image, Location location) {
+        this.userId = userId;
+        this.nickname = nickname;
+        this.email = email;
+        this.phoneNumber = phoneNumber;
+        this.image = image;
+        this.locationId = location;
+    }
 
 }
